*: relocate the cases
