+++ ddl/BUILD.bazel
@@ -198,7 +198,6 @@ go_test(
         "primary_key_handle_test.go",
         "reorg_partition_test.go",
         "repair_table_test.go",
-        "resource_group_test.go",
         "restart_test.go",
         "rollingback_test.go",
         "schema_test.go",
