+++ ddl/resourcegroup/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "resourcegroup",
@@ -12,18 +12,5 @@ go_library(
         "//parser/model",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_kvproto//pkg/resource_manager",
-    ],
-)
-
-go_test(
-    name = "resourcegroup_test",
-    timeout = "short",
-    srcs = ["group_test.go"],
-    embed = [":resourcegroup"],
-    flaky = True,
-    deps = [
-        "//parser/model",
-        "@com_github_pingcap_kvproto//pkg/resource_manager",
-        "@com_github_stretchr_testify//require",
     ],
 )
