+++ ddl/resourcegrouptest/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "resourcegrouptest_test",
+    srcs = ["resource_group_test.go"],
+    deps = [
+        "//ddl/internal/callback",
+        "//ddl/resourcegroup",
+        "//domain",
+        "//domain/infosync",
+        "//errno",
+        "//parser/model",
+        "//sessionctx",
+        "//testkit",
+        "@com_github_pingcap_kvproto//pkg/resource_manager",
+        "@com_github_stretchr_testify//require",
+    ],
+)
